Move `format_imports` and `format_import` methods to `imports.rs`
